requirements.txt
protobuf
tiktoken
sentencepiece
transformers[torch]
evaluate
seqeval
pytorch-crf
sklearn_crfsuite
